+++ services/netmanagernative/src/netsys/net_manager_native.cpp
@@ -79,7 +79,7 @@ int32_t NetManagerNative::NetworkCreatePhysical(int32_t netId, int32_t permissio
 
 int32_t NetManagerNative::NetworkCreateVirtual(int32_t netId, bool hasDns)
 {
-    return connManager_->CreateVirtualNetwork(static_cast<uint16_t>(netId), hasDns);
+    return connManager_->CreateVirtualNetwork(netId, hasDns);
 }
 
 int32_t NetManagerNative::NetworkDestroy(int32_t netId)
@@ -89,12 +89,12 @@ int32_t NetManagerNative::NetworkDestroy(int32_t netId)
 
 int32_t NetManagerNative::NetworkAddUids(int32_t netId, const std::vector<UidRange> &uidRanges)
 {
-    return connManager_->AddUidsToNetwork(static_cast<uint16_t>(netId), uidRanges);
+    return connManager_->AddUidsToNetwork(netId, uidRanges);
 }
 
 int32_t NetManagerNative::NetworkDelUids(int32_t netId, const std::vector<UidRange> &uidRanges)
 {
-    return connManager_->RemoveUidsFromNetwork(static_cast<uint16_t>(netId), uidRanges);
+    return connManager_->RemoveUidsFromNetwork(netId, uidRanges);
 }
 
 int32_t NetManagerNative::NetworkAddInterface(int32_t netId, std::string interfaceName)
